Removed a lot of junk files
A lot of files that are unused got removed.
There is one ancient file that has laid a curse upon this project. Do not disturb the file known as fix inside the custom character folder. For if you do it will unleash its wrath upon the game and turn most of the textures pink! You have been warned.